requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-docx>=1.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
pydantic>=2.6.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
aiohttp>=3.9.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Task Queue
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# WebSockets
websockets>=11.0.0

# Monitoring
prometheus-client>=0.17.0
structlog>=23.1.0

# Test dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
httpx>=0.26.0  # For TestClient